+++ ui/src/messages/networkGraph.js
@@ -1,0 +1,8 @@
+import { PROTOCOLS } from 'constants/networkGraph';
+
+const networkProtocolLabels = {
+    [PROTOCOLS.L4_PROTOCOL_TCP]: 'TCP',
+    [PROTOCOLS.L4_PROTOCOL_UDP]: 'UDP',
+};
+
+export default networkProtocolLabels;